mm: Skip initialization of struct pages freed later by HVO
The initialization of struct pages for gigantic pages is moved
from memmap_init_reserved_pages to gather_bootmem_prealloc as
the size of the gigantic page is not know during memmap_init_reserved_pages
(Could use PAGE_SIZE << (PUD_SHIFT - PAGE_SHIFT)??)

If the region is for hugepages and if HVO is enabled, then those
struct pages which will be freed later don't need to be initialized.
This can save significant time when a large number of hugepages are
allocated at boot time. As memmap_init_reserved_pages is only called at
boot time, we don't need to worry about memory hotplug.

Hugepage regions are kept separate from non hugepage regions in
memblock_merge_regions so that initialization for unused struct pages
can be skipped for the entire region.

Signed-off-by: Usama Arif <usama.arif@bytedance.com>