prctl: allow overriding system THP policy to always per process
PR_THP_POLICY_ALWAYS will allow the process to always get THPs
if the system global setting is "always" or "madvise" for hugepages
sizes that are set to "inherit". This won't work if system global is
set to "never".
PR_THP_POLICY_MADVISE will allow the process to get THPs on madvise
basis only if the system global setting is "always" or "madvise" for
hugepage sizes that are set to "inherit". This won't work if system
global is set to "never".

Signed-off-by: Usama Arif <usamaarif642@gmail.com>